[fix] Page 폴더 활용해서 App.jsx 라우팅 구조 리팩토링
